+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azuretools/ijidea/actions/AzureSignInAction.java
@@ -21,6 +21,9 @@
  */
 package com.microsoft.azuretools.ijidea.actions;
 
+import static com.microsoft.azuretools.telemetry.TelemetryConstants.ACCOUNT;
+import static com.microsoft.azuretools.telemetry.TelemetryConstants.SIGNOUT;
+
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.DataKeys;
 import com.intellij.openapi.diagnostic.Logger;
@@ -33,6 +36,8 @@ import com.microsoft.azuretools.authmanage.models.AuthMethodDetails;
 import com.microsoft.azuretools.ijidea.ui.ErrorWindow;
 import com.microsoft.azuretools.ijidea.ui.SignInWindow;
 import com.microsoft.azuretools.ijidea.utility.AzureAnAction;
+import com.microsoft.azuretools.telemetry.TelemetryConstants;
+import com.microsoft.azuretools.telemetrywrapper.EventUtil;
 import com.microsoft.intellij.helpers.UIHelperImpl;
 import com.microsoft.intellij.serviceexplorer.azure.SignInOutAction;
 import org.jetbrains.annotations.NotNull;
@@ -57,6 +62,16 @@ public class AzureSignInAction extends AzureAnAction {
     public void onActionPerformed(AnActionEvent e) {
         Project project = DataKeys.PROJECT.getData(e.getDataContext());
         onAzureSignIn(project);
+    }
+
+    @Override
+    protected String getServiceName() {
+        return ACCOUNT;
+    }
+
+    @Override
+    protected String getOperationName(AnActionEvent event) {
+        return TelemetryConstants.SIGNIN;
     }
 
     @Override
@@ -108,10 +123,12 @@ public class AzureSignInAction extends AzureAnAction {
                     JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE,
                     new ImageIcon("icons/azure.png"));
                 if (res == JOptionPane.OK_OPTION) {
-                    AdAuthManager adAuthManager = AdAuthManager.getInstance();
-                    if (adAuthManager.isSignedIn())
-                        adAuthManager.signOut();
-                    authMethodManager.signOut();
+                    EventUtil.executeWithLog(ACCOUNT, SIGNOUT, (operation) -> {
+                        AdAuthManager adAuthManager = AdAuthManager.getInstance();
+                        if (adAuthManager.isSignedIn())
+                            adAuthManager.signOut();
+                        authMethodManager.signOut();
+                    });
                 }
             } else {
                 doSignIn(authMethodManager, project);