+++ Utils/azure-explorer-common/src/com/microsoft/tooling/msservices/serviceexplorer/azure/vmarm/VMNode.java
@@ -21,6 +21,12 @@
  */
 package com.microsoft.tooling.msservices.serviceexplorer.azure.vmarm;
 
+import static com.microsoft.azuretools.telemetry.TelemetryConstants.DELETE_VM;
+import static com.microsoft.azuretools.telemetry.TelemetryConstants.RESTART_VM;
+import static com.microsoft.azuretools.telemetry.TelemetryConstants.SHUTDOWN_VM;
+import static com.microsoft.azuretools.telemetry.TelemetryConstants.START_VM;
+import static com.microsoft.azuretools.telemetry.TelemetryConstants.VM;
+
 import com.microsoft.azure.CloudException;
 import com.microsoft.azure.management.compute.InstanceViewStatus;
 import com.microsoft.azure.management.compute.VirtualMachine;
@@ -70,21 +76,29 @@ public class VMNode extends RefreshableNode implements TelemetryProperties {
                 if (azureManager == null) {
                     return;
                 }
-                azureManager.getAzure(subscriptionId).virtualMachines().deleteByResourceGroup(virtualMachine.resourceGroupName(), virtualMachine.name());
+                azureManager.getAzure(subscriptionId).virtualMachines().
+                    deleteByResourceGroup(virtualMachine.resourceGroupName(), virtualMachine.name());
             } catch (Exception ex) {
-
             }
-            DefaultLoader.getIdeHelper().invokeLater(new Runnable() {
-                @Override
-                public void run() {
-                    // instruct parent node to remove this node
-                    getParent().removeDirectChildNode(VMNode.this);
-                }
+
+            DefaultLoader.getIdeHelper().invokeLater(() -> {
+                // instruct parent node to remove this node
+                getParent().removeDirectChildNode(VMNode.this);
             });
         }
 
         @Override
         protected void onSubscriptionsChanged(NodeActionEvent e) throws AzureCmdException {
+        }
+
+        @Override
+        protected String getServiceName() {
+            return VM;
+        }
+
+        @Override
+        protected String getOperationName(NodeActionEvent event) {
+            return DELETE_VM;
         }
     }
 
@@ -105,6 +119,16 @@ public class VMNode extends RefreshableNode implements TelemetryProperties {
         @Override
         protected void onSubscriptionsChanged(NodeActionEvent e) throws AzureCmdException {
         }
+
+        @Override
+        protected String getServiceName() {
+            return VM;
+        }
+
+        @Override
+        protected String getOperationName(NodeActionEvent event) {
+            return RESTART_VM;
+        }
     }
 
     public class StartVMAction extends AzureNodeActionPromptListener {
@@ -123,6 +147,16 @@ public class VMNode extends RefreshableNode implements TelemetryProperties {
 
         @Override
         protected void onSubscriptionsChanged(NodeActionEvent e) throws AzureCmdException {
+        }
+
+        @Override
+        protected String getServiceName() {
+            return VM;
+        }
+
+        @Override
+        protected String getOperationName(NodeActionEvent event) {
+            return START_VM;
         }
     }
 
@@ -143,6 +177,16 @@ public class VMNode extends RefreshableNode implements TelemetryProperties {
 
         @Override
         protected void onSubscriptionsChanged(NodeActionEvent e) throws AzureCmdException {
+        }
+
+        @Override
+        protected String getServiceName() {
+            return VM;
+        }
+
+        @Override
+        protected String getOperationName(NodeActionEvent event) {
+            return SHUTDOWN_VM;
         }
     }