+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/sdk/cluster/ClusterDetail.java
@@ -32,15 +32,16 @@ import com.microsoft.azure.hdinsight.spark.common.SparkSubmitStorageTypeOptionsF
 import com.microsoft.azuretools.authmanage.models.SubscriptionDetail;
 import com.microsoft.azuretools.azurecommons.helpers.AzureCmdException;
 import com.microsoft.azuretools.azurecommons.helpers.Nullable;
+import com.microsoft.azuretools.telemetry.AppInsightsClient;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 
 import java.io.IOException;
 import java.net.URI;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, ILogger {
@@ -53,7 +54,6 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
     private final String DefaultFS = "fs.defaultFS";
     private final String FSDefaultName = "fs.default.name";
     private final String StorageAccountKeyPrefix = "fs.azure.account.key.";
-    private final String StorageAccountNamePattern = "^(abfs[s]?|wasb[s]?)://(.*)@(.*)$";
     private final String ResourceGroupStartTag = "resourceGroups/";
     private final String ResourceGroupEndTag = "/providers/";
 
@@ -69,6 +69,8 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
     private IHDIStorageAccount defaultStorageAccount;
     private List<HDStorageAccount> additionalStorageAccounts;
     private boolean isConfigInfoAvailable = false;
+    @Nullable
+    private Map<String, String> coresiteMap = null;
 
     public ClusterDetail(SubscriptionDetail paramSubscription,
                          ClusterRawInfo paramClusterRawInfo,
@@ -96,10 +98,23 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
 
     @Override
     public String getTitle() {
-        return Optional.ofNullable(getSparkVersion())
-                .filter(ver -> !ver.trim().isEmpty())
-                .map(ver -> getName() + " (Spark: " + ver + ")")
-                .orElse(getName());
+        StringBuilder titleStringBuilder = new StringBuilder(getName());
+
+        String sparkVersion = getSparkVersion();
+        if (StringUtils.isNotBlank(sparkVersion)) {
+            titleStringBuilder.append(String.format(" (Spark: %s)", sparkVersion));
+        }
+
+        if (ClusterManagerEx.getInstance().isHdiReaderCluster(this)) {
+            titleStringBuilder.append(" (Role: Reader)");
+        }
+
+        String state = getState();
+        if (StringUtils.isNotBlank(state) && !state.equalsIgnoreCase("Running")) {
+            titleStringBuilder.append(String.format(" (State: %s)", state));
+        }
+
+        return titleStringBuilder.toString();
     }
 
     @Override
@@ -262,6 +277,7 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
                         Map<String, String> coresSiteMap = configurations.getCoresite();
                         ClusterIdentity clusterIdentity = configurations.getClusterIdentity();
                         if (coresSiteMap != null) {
+                            this.coresiteMap = coresSiteMap;
                             this.defaultStorageAccount = getDefaultStorageAccount(coresSiteMap, clusterIdentity);
                             this.additionalStorageAccounts = getAdditionalStorageAccounts(coresSiteMap);
                         }
@@ -273,64 +289,112 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
         }
     }
 
-    private IHDIStorageAccount getDefaultStorageAccount(Map<String, String> coresiteMap, ClusterIdentity clusterIdentity) throws HDIException{
-        String containerAddress = null;
-        if(coresiteMap.containsKey(DefaultFS)){
-            containerAddress = coresiteMap.get(DefaultFS);
-        }else if(coresiteMap.containsKey(FSDefaultName)){
-            containerAddress = coresiteMap.get(FSDefaultName);
+    @Nullable
+    public String getDefaultStorageRootPath() {
+        log().info("Cluster ID: " + clusterRawInfo.getId());
+        Map<String, String> requestedCoresiteMap = null;
+
+        try {
+            if (!(clusterOperation instanceof ClusterOperationNewAPIImpl)) {
+                requestedCoresiteMap = this.coresiteMap;
+            } else {
+                requestedCoresiteMap =
+                        ((ClusterOperationNewAPIImpl) clusterOperation).getClusterCoreSiteRequest(clusterRawInfo.getId())
+                                .toBlocking()
+                                .singleOrDefault(null);
+            }
+
+            if (requestedCoresiteMap == null) {
+                log().warn("Error getting cluster core-site. coresiteMap is null.");
+                return null;
+            }
+        } catch (Exception ex) {
+            log().warn("Error getting cluster core-site. " + ExceptionUtils.getStackTrace(ex));
+            return null;
         }
 
-        if(containerAddress == null){
-            throw new HDIException("Failed to get default storage account");
+        String containerAddress = null;
+        if (requestedCoresiteMap.containsKey(DefaultFS)) {
+            containerAddress = requestedCoresiteMap.get(DefaultFS);
+        } else if (requestedCoresiteMap.containsKey(FSDefaultName)) {
+            containerAddress = requestedCoresiteMap.get(FSDefaultName);
+        } else {
+            log().warn("Error getting cluster default storage account. containerAddress is null.");
+            return null;
         }
 
         String scheme = URI.create(containerAddress).getScheme();
-
-        //for adls
-        if(ADL_HOME_PREFIX.equalsIgnoreCase(containerAddress)) {
+        if (ADL_HOME_PREFIX.equalsIgnoreCase(containerAddress)) {
             String accountName = "";
             String defaultRootPath = "";
 
-            if(coresiteMap.containsKey(ADLS_HOME_HOST_NAME)) {
-                accountName = coresiteMap.get(ADLS_HOME_HOST_NAME).split("\\.")[0];
+            if (requestedCoresiteMap.containsKey(ADLS_HOME_HOST_NAME)) {
+                accountName = requestedCoresiteMap.get(ADLS_HOME_HOST_NAME).split("\\.")[0];
             }
-            if(coresiteMap.containsKey(ADLS_HOME_MOUNTPOINT)) {
-                defaultRootPath = coresiteMap.get(ADLS_HOME_MOUNTPOINT);
-            }
-
-            URI rootURI = URI.create(String.format("%s://%s.azuredatalakestore.net", scheme, accountName)).resolve(defaultRootPath);
-            return new ADLSStorageAccount(this,true, clusterIdentity, rootURI);
-        } else if (Pattern.compile(StorageAccountNamePattern).matcher(containerAddress).matches()) {
-            String storageAccountName = getStorageAccountName(containerAddress);
-            if(storageAccountName == null){
-                throw new HDIException("Failed to get default storage account name");
+            if (requestedCoresiteMap.containsKey(ADLS_HOME_MOUNTPOINT)) {
+                defaultRootPath = requestedCoresiteMap.get(ADLS_HOME_MOUNTPOINT);
             }
 
-            String defaultContainerName = getDefaultContainerName(containerAddress);
-
-            String keyNameOfDefaultStorageAccountKey = StorageAccountKeyPrefix + storageAccountName;
-            String storageAccountKey = null;
-            if(coresiteMap.containsKey(keyNameOfDefaultStorageAccountKey)){
-                storageAccountKey = coresiteMap.get(keyNameOfDefaultStorageAccountKey);
-            }
-
-            if (!scheme.startsWith(ADLSGen2StorageAccount.DefaultScheme) && storageAccountKey == null) {
-                throw new HDIException("Failed to get default storage account key");
-            }
-
-            if (scheme.startsWith(ADLSGen2StorageAccount.DefaultScheme)) {
-                try {
-                    return new ADLSGen2StorageAccount(this, storageAccountName, storageAccountKey, true, defaultContainerName, scheme);
-                } catch (Exception ex) {
-                    throw new HDIException("Failed to init ADLSGEN2 account", ex);
-                }
-            } else {
-                return new HDStorageAccount(this, storageAccountName, storageAccountKey, true, defaultContainerName);
-            }
-
+            return URI.create(String.format("%s://%s.azuredatalakestore.net", scheme, accountName))
+                    .resolve(defaultRootPath)
+                    .toString();
+        } else if (Pattern.compile(StoragePathInfo.BlobPathPattern).matcher(containerAddress).matches()
+                || Pattern.compile(StoragePathInfo.AdlsGen2PathPattern).matcher(containerAddress).matches()) {
+            return containerAddress;
         } else {
+            final Map<String, String> properties = new HashMap<>();
+            properties.put("ErrorType", "Unknown HDInsight default storage type");
+            properties.put("coreSiteMap", StringUtils.join(requestedCoresiteMap));
+            properties.put("containerAddress", containerAddress);
+            properties.put("ClusterID", this.clusterRawInfo.getId());
+            AppInsightsClient.createByType(AppInsightsClient.EventType.Error, this.getClass().getSimpleName(), null, properties);
+
             return null;
+        }
+    }
+
+    @Nullable
+    private IHDIStorageAccount getDefaultStorageAccount(Map<String, String> coresiteMap, ClusterIdentity clusterIdentity) throws HDIException {
+        String defaultStorageRootPath = getDefaultStorageRootPath();
+        if (defaultStorageRootPath == null) {
+            throw new HDIException("Failed to get default storage root path");
+        }
+
+        StoragePathInfo pathInfo = new StoragePathInfo(defaultStorageRootPath);
+        switch (pathInfo.storageType) {
+            case ADLS:
+                return new ADLSStorageAccount(this, true, clusterIdentity, URI.create(defaultStorageRootPath));
+
+            case BLOB:
+                String storageAccountName = pathInfo.path.getHost();
+                if (StringUtils.isBlank(storageAccountName)) {
+                    throw new HDIException("Failed to get default storage account name");
+                }
+
+                String defaultContainerName = pathInfo.path.getUserInfo();
+                String defaultStorageAccountKey = StorageAccountKeyPrefix + storageAccountName;
+                String storageAccountKey = null;
+                if (coresiteMap.containsKey(defaultStorageAccountKey)) {
+                    storageAccountKey = coresiteMap.get(defaultStorageAccountKey);
+                }
+
+                if (storageAccountKey == null) {
+                   return null;
+                }
+
+                return new HDStorageAccount(this, storageAccountName, storageAccountKey, true, defaultContainerName);
+
+            case ADLSGen2:
+                String accountName = pathInfo.path.getHost();
+                if (StringUtils.isBlank(accountName)) {
+                    throw new HDIException("Failed to get default storage account name");
+                }
+
+                String fileSystem = pathInfo.path.getUserInfo();
+                return new ADLSGen2StorageAccount(this, accountName, null, true, fileSystem, pathInfo.path.getScheme());
+
+            default:
+                return null;
         }
     }
 
@@ -356,28 +420,6 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
         return storageAccounts;
     }
 
-    private String getStorageAccountName(String containerAddress){
-        Pattern r = Pattern.compile(StorageAccountNamePattern);
-        Matcher m = r.matcher(containerAddress);
-        if(m.find())
-        {
-            return m.group(3);
-        }
-
-        return null;
-    }
-
-    private String getDefaultContainerName(String containerAddress){
-        Pattern r = Pattern.compile(StorageAccountNamePattern);
-        Matcher m = r.matcher(containerAddress);
-        if(m.find())
-        {
-            return m.group(2);
-        }
-
-        return null;
-    }
-
     public String getLivyConnectionUrl() {
         return URI.create(getConnectionUrl()).resolve("livy/").toString();
     }
@@ -396,7 +438,7 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
 
     @Override
     public SparkSubmitStorageTypeOptionsForCluster getStorageOptionsType() {
-        StorageAccountTypeEnum type = StorageAccountTypeEnum.UNKNOWN;
+        StorageAccountType type = StorageAccountType.UNKNOWN;
 
         if (getStorageAccount() == null) {
             try {
@@ -411,11 +453,11 @@ public class ClusterDetail implements IClusterDetail, LivyCluster, YarnCluster, 
 
         if (isRoleTypeReader()) {
             return SparkSubmitStorageTypeOptionsForCluster.HDInsightReaderStorageTypeOptions;
-        } else if (type == StorageAccountTypeEnum.ADLS) {
+        } else if (type == StorageAccountType.ADLS) {
             return SparkSubmitStorageTypeOptionsForCluster.ClusterWithAdls;
-        } else if (type == StorageAccountTypeEnum.BLOB) {
+        } else if (type == StorageAccountType.BLOB) {
             return SparkSubmitStorageTypeOptionsForCluster.ClusterWithBlob;
-        } else if(type == StorageAccountTypeEnum.ADLSGen2){
+        } else if(type == StorageAccountType.ADLSGen2){
            return SparkSubmitStorageTypeOptionsForCluster.ClusterWithAdlsGen2;
         } else {
             return SparkSubmitStorageTypeOptionsForCluster.ClusterWithUnknown;